+++ feather-runtime/src/lib.rs
@@ -1,0 +1,7 @@
+#![warn(clippy::unwrap_used)]
+
+pub mod http;
+pub mod server;
+mod utils;
+
+pub use ::http::{HeaderMap, HeaderName, HeaderValue, Method, Uri};