+++ src/lib.rs
@@ -2,13 +2,11 @@
 
 /// The [`Middleware`] trait and some common middleware primitives.
 pub mod middleware;
-/// The [`Response`] type.
-mod response;
+
 /// Synchronous API for Feather.
 mod sync;
 
-pub use crate::middleware::Middleware;
-pub use crate::sync::AppConfig;
-pub use response::Response;
+pub use crate::middleware::{Middleware,MiddlewareResult};
+pub use feather_runtime::http::HttpRequest as Request;
+pub use feather_runtime::http::HttpResponse as Response;
 pub use sync::App;
-pub use tiny_http::Request;