+++ tests/test_app.rs
@@ -1,16 +1,15 @@
-use feather::Response;
 use feather::middleware::MiddlewareResult;
-use feather::{App, AppConfig};
+use feather::App;
+use feather::{Request, Response};
 use reqwest::blocking::get;
 use std::thread;
 
 #[test]
 fn test_app() {
-    let config = AppConfig { threads: 4 };
-    let mut app = App::new(config);
+    let mut app = App::new();
 
-    app.get("/", |_request: &mut _, response: &mut _| {
-        *response = Response::ok("Hello from Feather!");
+    app.get("/", |_request: &mut Request, response: &mut Response| {
+        response.send_text("Hello from Feather!");
         MiddlewareResult::Next
     });
     thread::spawn(move || {